+++ lib/src/serializers/language.dart
@@ -1,3 +1,0 @@
-enum Language { dart, java }
-
-enum CodeGenerationMode { client, server }